requirements.txt
# ── Visão Computacional ───────────────────────────────────────
opencv-python>=4.8.0
dlib==19.24.6                    # 20.x quebra no macOS arm64 (fp.h ausente)
face-recognition>=1.3.0          # usa dlib internamente
numpy>=1.24.0

# ── Web / API ─────────────────────────────────────────────────
flask>=3.0.0
requests>=2.31.0

# ── Utilitários ───────────────────────────────────────────────
python-dotenv>=1.0.0
pillow>=10.0.0                   # Para QR Code display

# ── WhatsApp QR Code (setup_whatsapp.py) ─────────────────────
qrcode>=7.4.2

# ── Twilio SMS (opcional — apenas se USE_TWILIO=true) ─────────
# twilio>=8.0.0
